requirements.txt
praw==7.3.0
discord.py==1.7.3
requests==2.25.1
discord==1.7.3
python-dotenv==0.18.0
youtube_dl==2021.6.6
